requirements.txt
mcp
anthropic
google-genai
smolagents[mcp]
python-dotenv
nest_asyncio
pydantic
pandas
ddgs
transformers